+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Utility/Texture2DAtlas.cs
@@ -147,7 +147,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                     TextureDimension.Tex2D,
                     false,
                     true,
-                    false);
+                    false,
+                    xrInstancing: true);
 
             m_AtlasAllocator = new AtlasAllocator(width, height);
         }