+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added HDRenderPipeline.RequestSkyEnvironmentUpdate function to force an update from script when sky is set to OnDemand
 - Added a Lighting and BackLighting slots in Lit, StackLit, Fabric and Hair master nodes
 - Added support for overriding terrain detail rendering shaders, via the render pipeline editor resources asset
+- Added xrInstancing flag support to RTHandle
 
 ### Fixed
 - Fixed logic to disable FPTL with stereo rendering